Moved the constants file to a new namespace
Moved constants for fulcrum view support
