+++ FulcrumInjector/App.xaml.cs
@@ -11,11 +11,11 @@ using FulcrumDriveService;
 using FulcrumEmailService;
 using FulcrumEncryption;
 using FulcrumInjector.FulcrumViewContent;
+using FulcrumInjector.FulcrumViewContent.FulcrumControls;
 using FulcrumInjector.FulcrumViewContent.FulcrumViewModels;
 using FulcrumInjector.FulcrumViewContent.FulcrumViewModels.InjectorCoreViewModels;
 using FulcrumInjector.FulcrumViewContent.FulcrumViews.InjectorCoreViews;
 using FulcrumInjector.FulcrumViewSupport;
-using FulcrumInjector.FulcrumViewSupport.FulcrumControls;
 using FulcrumInjector.FulcrumViewSupport.FulcrumModels.SettingsModels;
 using FulcrumInjector.FulcrumViewSupport.FulcrumStyles;
 using FulcrumJson;
