+++ FulcrumInjector/FulcrumViewContent/FulcrumViewModels/FulcrumPipeStatusViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows.Controls;
+using FulcrumInjector.FulcrumViewSupport;
 using FulcrumInjector.FulcrumViewSupport.FulcrumDataConverters;
 using SharpLogging;
 using SharpPipes;
