+++ FulcrumInjector/FulcrumViewContent/FulcrumViewModels/InjectorCoreViewModels/FulcrumInstalledHardwareViewModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Windows.Controls;
+using FulcrumInjector.FulcrumViewSupport;
 using FulcrumInjector.FulcrumViewSupport.FulcrumJsonSupport;
 using FulcrumJson;
 using SharpLogging;
