+++ FulcrumInjector/FulcrumViewContent/FulcrumViewModels/InjectorCoreViewModels/FulcrumSimulationPlaybackViewModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Windows.Controls;
+using FulcrumInjector.FulcrumViewSupport;
 using FulcrumInjector.FulcrumViewSupport.FulcrumJsonSupport;
 using FulcrumJson;
 using Newtonsoft.Json;
