+++ FulcrumInjector/FulcrumViewContent/FulcrumViews/InjectorCoreViews/FulcrumSimulationPlaybackView.xaml
@@ -10,7 +10,7 @@
              xmlns:fulcrumViewSupport="clr-namespace:FulcrumInjector.FulcrumViewSupport"
              xmlns:fulcrumDataConverters="clr-namespace:FulcrumInjector.FulcrumViewSupport.FulcrumDataConverters"
              xmlns:passThruTypes="clr-namespace:SharpWrapper.PassThruTypes;assembly=SharpWrapper"
-             xmlns:fulcrumControls="clr-namespace:FulcrumInjector.FulcrumViewSupport.FulcrumControls"
+             xmlns:fulcrumControls="clr-namespace:FulcrumInjector.FulcrumViewContent.FulcrumControls"
              mc:Ignorable="d" d:DesignHeight="450" d:DesignWidth="800" Loaded="FulcrumSimulationPlaybackView_OnLoaded"
              DataContext="{Binding RelativeSource={RelativeSource Mode=Self}, Path=ViewModel}">
