+++ FulcrumInjector/FulcrumViewContent/FulcrumViews/InjectorMiscViews/FulcrumUpdaterView.xaml.cs
@@ -7,6 +7,7 @@ using System.Windows.Documents;
 using System.Windows.Input;
 using System.Xaml;
 using FulcrumInjector.FulcrumViewContent.FulcrumViewModels.InjectorMiscViewModels;
+using FulcrumInjector.FulcrumViewSupport;
 using FulcrumInjector.FulcrumViewSupport.FulcrumDataConverters;
 using Markdig;
 using SharpLogging;
