+++ FulcrumInjector/FulcrumViewContent/FulcrumViews/InjectorOptionViews/FulcrumDebugLoggingView.xaml.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using FulcrumInjector.FulcrumViewContent.FulcrumViewModels.InjectorOptionViewModels;
+using FulcrumInjector.FulcrumViewSupport;
 using NLog;
 using SharpLogging;
